Clean up integration tests.
